# Conflicts:
#	app/src/main/java/com/example/seollyongdan_frontend/data/service/ApiKeyStorage.kt
#	app/src/main/java/com/example/seollyongdan_frontend/presentation/home/screen/HomeScreen.kt

+++ app/src/main/java/com/example/seollyongdan_frontend/presentation/community/screen/CommunitySearchScreen.kt
@@ -67,7 +67,7 @@ fun CommunitySearchRoute(
 @Composable
 fun CommunitySearchScreen(
     onBackClick: () -> Unit,
-    onDetailClick: (Long, String) -> Unit,
+    onDetailClick: (Int, String) -> Unit,
     district: String,
     searchViewModel: CommunitySearchViewModel
 ) {
